WIP: fixup desktop
