+++ editor/src/messages/frontend/frontend_message.rs
@@ -281,10 +281,6 @@ pub enum FrontendMessage {
 		layout_target: LayoutTarget,
 		diff: Vec<WidgetDiff>,
 	},
-	#[cfg(target_os = "macos")]
-	UpdateMenuBarLayoutForMac {
-		layout: Layout,
-	},
 	UpdateMouseCursor {
 		cursor: MouseCursorIcon,
 	},
